requirements.txt
selene==2.0.0rc9
pytest
allure-pytest
